+++ src/value/schemas/time.rs
@@ -1,6 +1,7 @@
 use crate::id::Id;
 use crate::id_hex;
 use crate::value::{FromValue, ToValue, Value, ValueSchema};
+use std::convert::Infallible;
 
 use std::convert::TryInto;
 
@@ -16,6 +17,7 @@ pub struct NsTAIInterval;
 
 impl ValueSchema for NsTAIInterval {
     const VALUE_SCHEMA_ID: Id = id_hex!("675A2E885B12FCBC0EEC01E6AEDD8AA8");
+    type ValidationError = Infallible;
 }
 
 impl ToValue<NsTAIInterval> for (Epoch, Epoch) {